separate into equations, diffusion_models, models, lit_modules
