+++ configs/ddim/cifar10.yaml
@@ -72,23 +72,6 @@ ckpt_path: null
 model:
   class_path: dmme.ddim.LitDDIM
   init_args:
-    model:
-      class_path: dmme.ddpm.UNet
-      init_args:
-        in_channels: 3
-        pos_dim: 128
-        emb_dim: 512
-        num_groups: 32
-        dropout: 0.1
-        channels_per_depth:
-          - 128
-          - 256
-          - 256
-          - 256
-        num_blocks: 2
-        attention_depths:
-          - 2
-    tau_schedule: "linear"
     lr: 2e-4
     warmup: 5000
     imgsize:
