--- src/dmme/ddim/__init__.py
@@ -1,2 +1,0 @@
-from .ddim import *
-from .litddim import *
