+++ src/dmme/diffusion_models/__init__.py
@@ -1,0 +1,4 @@
+from .ddpm import DDPM
+from .ddim import DDIM
+
+__all__ = ["DDPM", "DDIM"]
