+++ src/dmme/equations/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ["ddpm", "ddim"]
+
+from dmme.equations import ddpm
+from dmme.equations import ddim
