+++ src/dmme/equations/ddim/__init__.py
@@ -1,0 +1,3 @@
+from .ddim import linear_tau, quadratic_tau, reverse_process
+
+__all__ = ["linear_tau", "quadratic_tau", "reverse_process"]
