+++ src/dmme/lit_modules/__init__.py
@@ -1,0 +1,4 @@
+from .ddpm import LitDDPM
+from .ddim import LitDDIM
+
+__all__ = ["LitDDPM", "LitDDIM"]
