+++ src/dmme/models/__init__.py
@@ -1,0 +1,3 @@
+from .unet import UNet
+
+__all__ = ["UNet"]
