--- src/dmme/parsing.py
@@ -1,7 +1,0 @@
-def parse(pattern: str):
-    left, right = pattern.split("->")
-
-    inputs = left.split(" ")
-    output = right.strip()
-
-    return inputs, output
